split wiring pop-interaction controller & animator
At first I found it hard to follow why `-navigationController:animationControllerForOperation:fromViewController:toViewController:` was wiring the interaction controller to `toVC`.  After some fiddling in my own app, I found the real event you meant was when a view controller is pushed on the stack to prepare the pop transition.

I find the change makes the intent clear and the code easier to follow.